+++ changelog.md
@@ -70,6 +70,7 @@ Cleaned up the Power Sequencer menu based on @tintwotin's suggestions:
 ### Bug fixes ###
 
 - `Delete crossfade`: Fixed error when deleting a crossfade between color strips
+- `Add Crossfade`: Fixed not crossfading with color strips
 - `Expand to surrounding cuts`: Fixed breaking the strip if trying to use the tool on the the first or the last sequence in the scene.
 - `Playback Speed`: 
     - Fixed error when calling `Decrease Playback Speed`